backend/app/agents/medical_agents.py
from typing import Dict, List, Any
from datetime import datetime
import re
from langchain.agents import Tool, AgentExecutor, LLMSingleActionAgent
from langchain.prompts import StringPromptTemplate
from langchain_openai import ChatOpenAI
from langchain.schema import AgentAction, AgentFinish
import os
from dotenv import load_dotenv

load_dotenv()

class MedicalAgentSystem:
    def __init__(self):
        self.llm = ChatOpenAI(
            temperature=0,
            model="gpt-4",
            api_key=os.getenv("OPENAI_API_KEY")
        )
        
        # Initialize tools
        self.tools = [
            Tool(
                name="extract_dates",
                func=self._extract_dates,
                description="Extract dates from medical text"
            ),
            Tool(
                name="identify_conditions",
                func=self._identify_conditions,
                description="Identify medical conditions from text"
            ),
            Tool(
                name="generate_summary",
                func=self._generate_summary,
                description="Generate a simplified summary of medical text"
            ),
            Tool(
                name="find_resources",
                func=self._find_resources,
                description="Find relevant medical resources for conditions"
            )
        ]
    
    async def process_document(self, text: str) -> Dict[str, Any]:
        """
        Process medical document using the agent system
        """
        try:
            # Extract dates
            dates = self._extract_dates(text)
            
            # Identify conditions
            conditions = self._identify_conditions(text)
            
            # Generate summary
            summary = self._generate_summary(text)
            
            # Find resources
            resources = self._find_resources(conditions)
            
            return {
                "dates": dates,
                "conditions": conditions,
                "summary": summary,
                "resources": resources,
                "processed_date": datetime.now().isoformat()
            }
            
        except Exception as e:
            raise Exception(f"Error in agent processing: {str(e)}")
    
    def _extract_dates(self, text: str) -> List[str]:
        """
        Extract dates from medical text
        """
        # Simple date pattern matching
        date_pattern = r'\d{1,2}[-/]\d{1,2}[-/]\d{2,4}'
        dates = re.findall(date_pattern, text)
        return dates
    
    def _identify_conditions(self, text: str) -> List[str]:
        """
        Identify medical conditions from text
        """
        # TODO: Implement more sophisticated condition identification
        # This could use a medical NER model or a comprehensive medical terms database
        return []
    
    def _generate_summary(self, text: str) -> str:
        """
        Generate a simplified summary of medical text
        """
        # TODO: Implement more sophisticated summarization
        # This could use GPT-4 with specific medical summarization prompts
        return "Summary generation not yet implemented"
    
    def _find_resources(self, conditions: List[str]) -> List[Dict[str, str]]:
        """
        Find relevant medical resources for conditions
        """
        # TODO: Implement resource finding
        # This could integrate with medical databases and APIs
        return [] 