backend/app/processors/document_processor.py
from PIL import Image
import easyocr
from pathlib import Path
import json
from typing import Dict, Any
import os
from datetime import datetime
import numpy as np

class DocumentProcessor:
    def __init__(self):
        # Initialize EasyOCR reader
        # Using English and medical terms
        self.reader = easyocr.Reader(['en'])
    
    async def process_document(self, file_path: Path) -> Dict[str, Any]:
        """
        Process a document through the pipeline:
        1. Extract text using OCR
        2. Process with agents
        3. Generate summary and resources
        """
        try:
            # Extract text using OCR
            text = self._extract_text(file_path)
            
            # Process the extracted text
            processed_data = await self._process_text(text)
            
            return processed_data
            
        except Exception as e:
            raise Exception(f"Error processing document: {str(e)}")
    
    def _extract_text(self, file_path: Path) -> str:
        """
        Extract text from image using EasyOCR
        """
        try:
            # Read the image
            image = Image.open(file_path)
            
            # Convert PIL Image to numpy array for EasyOCR
            image_np = np.array(image)
            
            # Extract text using EasyOCR
            results = self.reader.readtext(image_np)
            
            # Combine all detected text
            text = '\n'.join([result[1] for result in results])
            
            return text.strip()
            
        except Exception as e:
            raise Exception(f"Error extracting text: {str(e)}")
    
    async def _process_text(self, text: str) -> Dict[str, Any]:
        """
        Process the extracted text with agents
        """
        # TODO: Implement agent-based processing
        # This will include:
        # 1. Date extraction
        # 2. Condition identification
        # 3. Summary generation
        # 4. Resource gathering
        
        # Placeholder response
        return {
            "original_text": text,
            "processed_date": datetime.now().isoformat(),
            "summary": "Document processing not yet implemented",
            "conditions": [],
            "resources": []
        } 