backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.27
pydantic==2.6.1
python-dotenv==1.0.1
pillow==10.2.0
easyocr==1.7.1
langchain==0.1.5
openai==1.12.0
python-magic==0.4.27
aiofiles==23.2.1 